+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added prometheus extension
+- Added `in!timezone` command for changing the timezone of incidents
+
+### Changed
+
+- Changed to urlsafe base64 for error logs
 
 
 ## [0.1.1]